+++ src/Transformers/IntervalDiscretizer.php
@@ -47,34 +47,9 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
     /**
      * The bin intervals of the fitted data.
      *
-     * @var array<(int|float)[]>|null
+     * @var array<float[]>|null
      */
     protected ?array $intervals = null;
-
-    /**
-     * Convert an integer to a base 26 string.
-     *
-     * @param int $value
-     * @return string
-     */
-    protected static function base26(int $value) : string
-    {
-        if ($value < 0) {
-            return '-' . self::base26(-$value);
-        }
-
-        $base26 = '';
-
-        while ($value >= 0) {
-            $base26 = chr(97 + $value % 26) . $base26;
-
-            $value = intdiv($value, 26);
-
-            --$value;
-        }
-
-        return $base26;
-    }
 
     /**
      * @param int $bins
@@ -182,7 +157,7 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
 
                 foreach ($interval as $ordinal => $edge) {
                     if ($value <= $edge) {
-                        $value = self::base26($ordinal);
+                        $value = $ordinal;
 
                         break;
                     }