+++ tests/Transformers/IntervalDiscretizerTest.php
@@ -51,7 +51,7 @@ class IntervalDiscretizerTest extends TestCase
 
         $this->assertCount(4, $sample);
 
-        $expected = ['a', 'b', 'c', 'd', 'e'];
+        $expected = [0, 1, 2, 3, 4];
 
         $this->assertContains($sample[0], $expected);
         $this->assertContains($sample[1], $expected);